fixing visibilities between modules
