+++ core/data/build.gradle.kts
@@ -48,6 +48,7 @@ android {
 }
 
 dependencies {
+  implementation(project(":core:domain"))
 
   implementation(libs.androidx.core.ktx)
   implementation(libs.androidx.appcompat)
