--- core/data/src/main/java/com/github/pepitoria/blinkoapp/data/model/login/LoginRequest.kt
@@ -1,4 +1,0 @@
-package com.github.pepitoria.blinkoapp.data.model.login
-
-class LoginRequest {
-}
