+++ core/data/src/main/java/com/github/pepitoria/blinkoapp/data/repository/auth/AuthenticationRepositoryApiImpl.kt
@@ -1,9 +1,8 @@
 package com.github.pepitoria.blinkoapp.data.repository.auth
 
-import com.github.pepitoria.blinkoapp.data.localstorage.LocalStorage
-import com.github.pepitoria.blinkoapp.data.model.login.LoginRequest
-import com.github.pepitoria.blinkoapp.data.model.login.LoginResponse
-import com.github.pepitoria.blinkoapp.data.model.session.SessionDto
+import com.github.pepitoria.blinkoapp.domain.data.LocalStorage
+import com.github.pepitoria.blinkoapp.domain.data.model.session.SessionDto
+import com.github.pepitoria.blinkoapp.domain.data.AuthenticationRepository
 import javax.inject.Inject
 
 private const val URL_KEY = "com.github.pepitoria.blinkoapp.data.repository.auth.URL_KEY"
@@ -12,10 +11,6 @@ private const val TOKEN_KEY = "com.github.pepitoria.blinkoapp.data.repository.au
 class AuthenticationRepositoryApiImpl @Inject constructor(
   private val localStorage: LocalStorage
 ) : AuthenticationRepository {
-  override suspend fun login(loginRequest: LoginRequest): LoginResponse {
-    //TODO: no api call yet
-    return LoginResponse()
-  }
 
   override fun saveSession(sessionDto: SessionDto) {
     localStorage.saveString(URL_KEY, sessionDto.url)
