+++ core/domain/src/main/java/com/github/pepitoria/blinkoapp/domain/LocalStorageUseCases.kt
@@ -1,11 +1,6 @@
 package com.github.pepitoria.blinkoapp.domain
 
-import com.github.pepitoria.blinkoapp.data.localstorage.LocalStorage
-import com.github.pepitoria.blinkoapp.data.model.login.LoginRequest
-import com.github.pepitoria.blinkoapp.data.model.notelist.NoteListRequest
-import com.github.pepitoria.blinkoapp.data.repository.auth.AuthenticationRepository
-import com.github.pepitoria.blinkoapp.data.repository.note.NoteRepository
-import timber.log.Timber
+import com.github.pepitoria.blinkoapp.domain.data.LocalStorage
 import javax.inject.Inject
 
 class LocalStorageUseCases @Inject constructor(
