+++ core/domain/src/main/java/com/github/pepitoria/blinkoapp/domain/data/AuthenticationRepository.kt
@@ -1,0 +1,10 @@
+package com.github.pepitoria.blinkoapp.domain.data
+
+import com.github.pepitoria.blinkoapp.domain.data.model.session.SessionDto
+
+
+interface AuthenticationRepository {
+    fun saveSession(sessionDto: SessionDto)
+    fun getSession(): SessionDto?
+    fun logout()
+}
