+++ core/domain/src/main/java/com/github/pepitoria/blinkoapp/domain/data/model/ApiResult.kt
@@ -1,4 +1,4 @@
-package com.github.pepitoria.blinkoapp.data.model
+package com.github.pepitoria.blinkoapp.domain.data.model
 
 sealed class ApiResult<out T> {
     data class ApiSuccess<out R>(val value: R) : ApiResult<R>()
