+++ core/domain/src/main/java/com/github/pepitoria/blinkoapp/domain/data/model/login/LoginRequest.kt
@@ -1,0 +1,4 @@
+package com.github.pepitoria.blinkoapp.domain.data.model.login
+
+class LoginRequest {
+}
