+++ core/domain/src/main/java/com/github/pepitoria/blinkoapp/domain/data/model/notelist/NoteResponse.kt
@@ -1,4 +1,4 @@
-package com.github.pepitoria.blinkoapp.data.model.notelist
+package com.github.pepitoria.blinkoapp.domain.data.model.notelist
 
 import com.google.gson.annotations.SerializedName
