+++ core/domain/src/main/java/com/github/pepitoria/blinkoapp/domain/data/model/session/SessionDto.kt
@@ -1,4 +1,4 @@
-package com.github.pepitoria.blinkoapp.data.model.session
+package com.github.pepitoria.blinkoapp.domain.data.model.session
 
 data class SessionDto(
   val url: String,
